requirements.txt
Flask==3.0.0
openpyxl==3.1.2
python-docx==1.1.0
gunicorn==21.2.0
Werkzeug==3.0.0